General improvements

+++ src/main/java/com/gmt/gmttechchallenge/api/videostatistics/VideosStatisticsPerSourceResponse.java
@@ -2,5 +2,5 @@ package com.gmt.gmttechchallenge.api.videostatistics;
 
 import com.gmt.gmttechchallenge.domain.VideoSource;
 
-public record VideosStatisticsPerSourceResponse(VideoSource source, int totalVideosImported, double averageDuration) {
+public record VideosStatisticsPerSourceResponse(VideoSource source, int totalVideosImported, long averageDuration) {
 }
